Make itsyouonline integration more modular
