+++ conf/app.ini
@@ -73,16 +73,6 @@ ORG_PAGING_NUM = 50
 ; Number of repos that are showed in one page
 REPO_PAGING_NUM = 15
 
-[itsyouonline]
-CLIENT_ID     = itsyouref
-CLIENT_SECRET = khZNlrGFiVqHb9u7h6Kh5IvZXY_aE1gWYL_v2Ike9WOZkza4j2k9
-REDIRECT_URL  = http://localhost:3000/oauth/redirect
-AUTH_URL      = https://itsyou.online/v1/oauth/authorize
-TOKEN_URL     = https://itsyou.online/v1/oauth/access_token
-SCOPE        = user:email
-
-
-
 [markdown]
 ; Enable hard line break extension
 ENABLE_HARD_LINE_BREAK = false
