+++ public/less/_iyo.less
@@ -1,0 +1,68 @@
+body {
+  font-family: "Helvetica Neue", "Meiryo Ui", "Microsoft YaHei", Arial, Helvetica, sans-serif !important;
+}
+
+.following.bar {
+  .head.link.item {
+		.dropdown.icon,
+		.menu .octicon {
+			margin-right: 5px;
+		}
+	}
+}
+
+// ensure Semantic UI .icon font family and 14px size does not apply to .icon.octicon
+.octicon.icon,
+.mega-octicon.icon {
+	font-family: octicons;
+	font-size: 16px;
+	opacity: 1 !important;
+}
+
+.following.bar.light{
+	background-color: #34495E;
+}
+.following.bar .brand .ui.mini.image{
+	width: 110px;
+}
+
+.following.bar .top.menu a.item, .ui.menu .item{
+	color: #fff;
+}
+
+.following.bar .top.menu a.item:hover, .ui.secondary.menu .active.item{
+	color: #fff;
+	background-color: rgb(39, 58, 76);
+}
+
+.ui.menu .item{
+	color: #C3C3C3;
+}
+.ui.secondary.menu .dropdown.item:hover, .ui.secondary.menu .link.item:hover, .ui.secondary.menu a.item:hover{
+	color: #fff;
+}
+
+.home{
+	padding-top: 60px;
+}
+
+.home .hero .octicon{
+	color: #34495E!important;
+}
+
+.home a{
+	color: #8AB10B!important;
+}
+
+.ui.blue.button, .ui.blue.buttons .button{
+	background-color: #34495E;
+}
+
+.ui.blue.button:hover, .ui.blue.buttons .button:hover{
+	background-color: #223140;
+}
+
+.ui.attached.header .right .button{
+	padding: 6px 10px!important;
+	font-size: 16px;
+}
