+++ public/less/gogs.less
@@ -11,3 +11,4 @@
 @import "_dashboard";
 @import "_admin";
 @import "_explore";
+@import "_iyo";
