+++ routers/api/v1/repo/collaborators.go
@@ -36,20 +36,3 @@ func AddCollaborator(ctx *context.APIContext, form api.AddCollaboratorOption) {
 
 	ctx.Status(204)
 }
-
-func DeleteCollaborator(ctx *context.APIContext, form api.AddCollaboratorOption) {
-	collaborator, err := models.GetUserByName(ctx.Params(":collaborator"))
-	if err != nil {
-		if models.IsErrUserNotExist(err) {
-			ctx.Error(422, "", err)
-		} else {
-			ctx.Error(500, "GetUserByName", err)
-		}
-		return
-	}
-
-	if err = ctx.Repo.Repository.DeleteCollaboration(collaborator.ID); err != nil {
-		ctx.Status(204)
-	}
-
-}
